Refactor into packages
