+++ src/main/java/duke/exception/DukeLoadingException.java
@@ -1,9 +1,11 @@
+package duke.exception;
+
 /**
  * Signals when storage cannot be properly loaded
  */
 public class DukeLoadingException extends Exception {
     /**
-     * Constructs an DukeLoadingException with the specified detail message.
+     * Constructs an duke.exception.DukeLoadingException with the specified detail message.
      *
      * @param e - description of the bad input
      */
